testing framework - renamed step methods
