+++ framework/scenario/src/facade.rs
@@ -3,6 +3,7 @@ mod debugger_backend;
 mod scenario_world;
 mod scenario_world_runner;
 mod scenario_world_steps;
+mod scenario_world_steps_deprecated;
 mod scenario_world_whitebox;
 mod whitebox_contract;
